pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nipux"
version = "0.1.0"
description = "A barebones local-model browser worker that runs restartable background jobs"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Nipux" }]
license = { text = "MIT" }
dependencies = [
  "openai>=2.21.0,<3",
  "pyyaml>=6.0.2,<7",
]

[project.optional-dependencies]
dev = ["pytest>=9.0.2,<10", "ruff"]

[project.scripts]
nipux = "nipux_cli.cli:main"

[tool.setuptools.packages.find]
include = ["nipux_cli", "nipux_cli.*"]

[tool.pytest.ini_options]
testpaths = ["tests/nipux_cli"]
addopts = "-q"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.uv]
exclude-newer = "7 days"
